Fix cast
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/dbuskit/trunk@31180 72102866-910b-0410-8b05-ffd578937521

+++ Source/DKArgument.m
@@ -804,7 +804,7 @@ DKDBusTypeForUnboxingObject(id object)
        IMP unboxFun = [value methodForSelector: aSelector];
 
        // Cast to void* first so that we don't get any funny implicit casts
-       *buffer = (long long)(void*)unboxFun(value, aSelector);
+       *buffer = (long long)(uintptr_t)(void*)unboxFun(value, aSelector);
        return YES;
      }
    }
